split test_event_pipeline, wip never modify event value
