+++ inputremapper/configs/input_config.py
@@ -21,6 +21,7 @@ from __future__ import annotations
 
 import itertools
 from typing import Tuple, Iterable, Union, List, Dict, Optional, Hashable
+import traceback
 
 from evdev import ecodes
 from inputremapper.configs.paths import PathUtils
@@ -115,6 +116,7 @@ class InputConfig(BaseModel):
     def from_input_event(cls, event: InputEvent) -> InputConfig:
         """create an input confing from the given InputEvent, uses the value as
         analog threshold"""
+        print("AAA from_input_event", event, event.value)
         return cls(
             type=event.type,
             code=event.code,
@@ -267,6 +269,11 @@ class InputConfig(BaseModel):
         analog_threshold: Optional[int] = None,
     ) -> InputConfig:
         """Return a new modified event."""
+        print("AAA modify", (
+                analog_threshold
+                if analog_threshold is not None
+                else self.analog_threshold
+            ))
         return InputConfig(
             type=type_ if type_ is not None else self.type,
             code=code if code is not None else self.code,
@@ -398,6 +405,7 @@ class InputCombination(Tuple[InputConfig, ...]):
         dicts = []
         for tuple_ in tuples:
             if len(tuple_) == 3:
+                print("AAA from_tuples", tuple_[2])
                 dicts.append(
                     {
                         "type": tuple_[0],
