+++ inputremapper/gui/reader_client.py
@@ -194,6 +194,7 @@ class ReaderClient:
 
     @staticmethod
     def _input_event_to_config(event: InputEvent):
+        print("AAA _input_event_to_config", event, event.value)
         return {
             "type": event.type,
             "code": event.code,
